Crate rusoto_inspector

Amazon Inspector
Amazon Inspector enables you to analyze the behavior of your AWS resources and to identify potential security issues. For more information, see Amazon Inspector User Guide.

If you're using the service, you're probably looking for InspectorClient and Inspector.

Structs

AddAttributesToFindingsRequest

AddAttributesToFindingsResponse

AgentAlreadyRunningAssessment
Used in the exception error that is thrown if you start an assessment run for an assessment target that includes an EC2 instance that is already participating in another started assessment run.

AgentFilter
Contains information about an Amazon Inspector agent. This data type is used as a request parameter in the ListAssessmentRunAgents action.

AgentPreview
Used as a response element in the PreviewAgents action.

AssessmentRun
A snapshot of an Amazon Inspector assessment run that contains the findings of the assessment run .

Used as the response element in the DescribeAssessmentRuns action.

AssessmentRunAgent
Contains information about an Amazon Inspector agent. This data type is used as a response element in the ListAssessmentRunAgents action.

AssessmentRunFilter
Used as the request parameter in the ListAssessmentRuns action.

AssessmentRunNotification
Used as one of the elements of the AssessmentRun data type.

AssessmentRunStateChange
Used as one of the elements of the AssessmentRun data type.

AssessmentTarget
Contains information about an Amazon Inspector application. This data type is used as the response element in the DescribeAssessmentTargets action.

AssessmentTargetFilter
Used as the request parameter in the ListAssessmentTargets action.

AssessmentTemplate
Contains information about an Amazon Inspector assessment template. This data type is used as the response element in the DescribeAssessmentTemplates action.

AssessmentTemplateFilter
Used as the request parameter in the ListAssessmentTemplates action.

AssetAttributes
A collection of attributes of the host from which the finding is generated.

Attribute
This data type is used as a request parameter in the AddAttributesToFindings and CreateAssessmentTemplate actions.

CreateAssessmentTargetRequest

CreateAssessmentTargetResponse

CreateAssessmentTemplateRequest

CreateAssessmentTemplateResponse

CreateExclusionsPreviewRequest

CreateExclusionsPreviewResponse

CreateResourceGroupRequest

CreateResourceGroupResponse

DeleteAssessmentRunRequest

DeleteAssessmentTargetRequest

DeleteAssessmentTemplateRequest

DescribeAssessmentRunsRequest

DescribeAssessmentRunsResponse

DescribeAssessmentTargetsRequest

DescribeAssessmentTargetsResponse

DescribeAssessmentTemplatesRequest

DescribeAssessmentTemplatesResponse

DescribeCrossAccountAccessRoleResponse

DescribeExclusionsRequest

DescribeExclusionsResponse

DescribeFindingsRequest

DescribeFindingsResponse

DescribeResourceGroupsRequest

DescribeResourceGroupsResponse

DescribeRulesPackagesRequest

DescribeRulesPackagesResponse

DurationRange
This data type is used in the AssessmentTemplateFilter data type.

EventSubscription
This data type is used in the Subscription data type.

Exclusion
Contains information about what was excluded from an assessment run.

ExclusionPreview
Contains information about what is excluded from an assessment run given the current state of the assessment template.

FailedItemDetails
Includes details about the failed items.

Finding
Contains information about an Amazon Inspector finding. This data type is used as the response element in the DescribeFindings action.

FindingFilter
This data type is used as a request parameter in the ListFindings action.

GetAssessmentReportRequest

GetAssessmentReportResponse

GetExclusionsPreviewRequest

GetExclusionsPreviewResponse

GetTelemetryMetadataRequest

GetTelemetryMetadataResponse

InspectorClient
A client for the Amazon Inspector API.

InspectorServiceAttributes
This data type is used in the Finding data type.

ListAssessmentRunAgentsRequest

ListAssessmentRunAgentsResponse

ListAssessmentRunsRequest

ListAssessmentRunsResponse

ListAssessmentTargetsRequest

ListAssessmentTargetsResponse

ListAssessmentTemplatesRequest

ListAssessmentTemplatesResponse

ListEventSubscriptionsRequest

ListEventSubscriptionsResponse

ListExclusionsRequest

ListExclusionsResponse

ListFindingsRequest

ListFindingsResponse

ListRulesPackagesRequest

ListRulesPackagesResponse

ListTagsForResourceRequest

ListTagsForResourceResponse

NetworkInterface
Contains information about the network interfaces interacting with an EC2 instance. This data type is used as one of the elements of the AssetAttributes data type.

PreviewAgentsRequest

PreviewAgentsResponse

PrivateIp
Contains information about a private IP address associated with a network interface. This data type is used as a response element in the DescribeFindings action.

RegisterCrossAccountAccessRoleRequest

RemoveAttributesFromFindingsRequest

RemoveAttributesFromFindingsResponse

ResourceGroup
Contains information about a resource group. The resource group defines a set of tags that, when queried, identify the AWS resources that make up the assessment target. This data type is used as the response element in the DescribeResourceGroups action.

ResourceGroupTag
This data type is used as one of the elements of the ResourceGroup data type.

RulesPackage
Contains information about an Amazon Inspector rules package. This data type is used as the response element in the DescribeRulesPackages action.

Scope
This data type contains key-value pairs that identify various Amazon resources.

SecurityGroup
Contains information about a security group associated with a network interface. This data type is used as one of the elements of the NetworkInterface data type.

SetTagsForResourceRequest

StartAssessmentRunRequest

StartAssessmentRunResponse

StopAssessmentRunRequest

SubscribeToEventRequest

Subscription
This data type is used as a response element in the ListEventSubscriptions action.

Tag
A key and value pair. This data type is used as a request parameter in the SetTagsForResource action and a response element in the ListTagsForResource action.

TelemetryMetadata
The metadata about the Amazon Inspector application data metrics collected by the agent. This data type is used as the response element in the GetTelemetryMetadata action.

TimestampRange
This data type is used in the AssessmentRunFilter data type.

UnsubscribeFromEventRequest

UpdateAssessmentTargetRequest

Enums

AddAttributesToFindingsError
Errors returned by AddAttributesToFindings

CreateAssessmentTargetError
Errors returned by CreateAssessmentTarget

CreateAssessmentTemplateError
Errors returned by CreateAssessmentTemplate

CreateExclusionsPreviewError
Errors returned by CreateExclusionsPreview

CreateResourceGroupError
Errors returned by CreateResourceGroup

DeleteAssessmentRunError
Errors returned by DeleteAssessmentRun

DeleteAssessmentTargetError
Errors returned by DeleteAssessmentTarget

DeleteAssessmentTemplateError
Errors returned by DeleteAssessmentTemplate

DescribeAssessmentRunsError
Errors returned by DescribeAssessmentRuns

DescribeAssessmentTargetsError
Errors returned by DescribeAssessmentTargets

DescribeAssessmentTemplatesError
Errors returned by DescribeAssessmentTemplates

DescribeCrossAccountAccessRoleError
Errors returned by DescribeCrossAccountAccessRole

DescribeExclusionsError
Errors returned by DescribeExclusions

DescribeFindingsError
Errors returned by DescribeFindings

DescribeResourceGroupsError
Errors returned by DescribeResourceGroups

DescribeRulesPackagesError
Errors returned by DescribeRulesPackages

GetAssessmentReportError
Errors returned by GetAssessmentReport

GetExclusionsPreviewError
Errors returned by GetExclusionsPreview

GetTelemetryMetadataError
Errors returned by GetTelemetryMetadata

ListAssessmentRunAgentsError
Errors returned by ListAssessmentRunAgents

ListAssessmentRunsError
Errors returned by ListAssessmentRuns

ListAssessmentTargetsError
Errors returned by ListAssessmentTargets

ListAssessmentTemplatesError
Errors returned by ListAssessmentTemplates

ListEventSubscriptionsError
Errors returned by ListEventSubscriptions

ListExclusionsError
Errors returned by ListExclusions

ListFindingsError
Errors returned by ListFindings

ListRulesPackagesError
Errors returned by ListRulesPackages

ListTagsForResourceError
Errors returned by ListTagsForResource

PreviewAgentsError
Errors returned by PreviewAgents

RegisterCrossAccountAccessRoleError
Errors returned by RegisterCrossAccountAccessRole

RemoveAttributesFromFindingsError
Errors returned by RemoveAttributesFromFindings

SetTagsForResourceError
Errors returned by SetTagsForResource

StartAssessmentRunError
Errors returned by StartAssessmentRun

StopAssessmentRunError
Errors returned by StopAssessmentRun

SubscribeToEventError
Errors returned by SubscribeToEvent

UnsubscribeFromEventError
Errors returned by UnsubscribeFromEvent

UpdateAssessmentTargetError
Errors returned by UpdateAssessmentTarget

Traits

Inspector
Trait representing the capabilities of the Amazon Inspector API. Amazon Inspector clients implement this trait.