Trait rusoto_inspector::Inspector

pub trait Inspector {
    fn add_attributes_to_findings(
        &self, 
        input: AddAttributesToFindingsRequest
    ) -> RusotoFuture<AddAttributesToFindingsResponse, AddAttributesToFindingsError>;

    fn create_assessment_target(
        &self, 
        input: CreateAssessmentTargetRequest
    ) -> RusotoFuture<CreateAssessmentTargetResponse, CreateAssessmentTargetError>;

    fn create_assessment_template(
        &self, 
        input: CreateAssessmentTemplateRequest
    ) -> RusotoFuture<CreateAssessmentTemplateResponse, CreateAssessmentTemplateError>;

    fn create_exclusions_preview(
        &self, 
        input: CreateExclusionsPreviewRequest
    ) -> RusotoFuture<CreateExclusionsPreviewResponse, CreateExclusionsPreviewError>;

    fn create_resource_group(
        &self, 
        input: CreateResourceGroupRequest
    ) -> RusotoFuture<CreateResourceGroupResponse, CreateResourceGroupError>;

    fn delete_assessment_run(
        &self, 
        input: DeleteAssessmentRunRequest
    ) -> RusotoFuture<(), DeleteAssessmentRunError>;

    fn delete_assessment_target(
        &self, 
        input: DeleteAssessmentTargetRequest
    ) -> RusotoFuture<(), DeleteAssessmentTargetError>;

    fn delete_assessment_template(
        &self, 
        input: DeleteAssessmentTemplateRequest
    ) -> RusotoFuture<(), DeleteAssessmentTemplateError>;

    fn describe_assessment_runs(
        &self, 
        input: DescribeAssessmentRunsRequest
    ) -> RusotoFuture<DescribeAssessmentRunsResponse, DescribeAssessmentRunsError>;

    fn describe_assessment_targets(
        &self, 
        input: DescribeAssessmentTargetsRequest
    ) -> RusotoFuture<DescribeAssessmentTargetsResponse, DescribeAssessmentTargetsError>;

    fn describe_assessment_templates(
        &self, 
        input: DescribeAssessmentTemplatesRequest
    ) -> RusotoFuture<DescribeAssessmentTemplatesResponse, DescribeAssessmentTemplatesError>;

    fn describe_cross_account_access_role(
        &self
    ) -> RusotoFuture<DescribeCrossAccountAccessRoleResponse, DescribeCrossAccountAccessRoleError>;

    fn describe_exclusions(
        &self, 
        input: DescribeExclusionsRequest
    ) -> RusotoFuture<DescribeExclusionsResponse, DescribeExclusionsError>;

    fn describe_findings(
        &self, 
        input: DescribeFindingsRequest
    ) -> RusotoFuture<DescribeFindingsResponse, DescribeFindingsError>;

    fn describe_resource_groups(
        &self, 
        input: DescribeResourceGroupsRequest
    ) -> RusotoFuture<DescribeResourceGroupsResponse, DescribeResourceGroupsError>;

    fn describe_rules_packages(
        &self, 
        input: DescribeRulesPackagesRequest
    ) -> RusotoFuture<DescribeRulesPackagesResponse, DescribeRulesPackagesError>;

    fn get_assessment_report(
        &self, 
        input: GetAssessmentReportRequest
    ) -> RusotoFuture<GetAssessmentReportResponse, GetAssessmentReportError>;

    fn get_exclusions_preview(
        &self, 
        input: GetExclusionsPreviewRequest
    ) -> RusotoFuture<GetExclusionsPreviewResponse, GetExclusionsPreviewError>;

    fn get_telemetry_metadata(
        &self, 
        input: GetTelemetryMetadataRequest
    ) -> RusotoFuture<GetTelemetryMetadataResponse, GetTelemetryMetadataError>;

    fn list_assessment_run_agents(
        &self, 
        input: ListAssessmentRunAgentsRequest
    ) -> RusotoFuture<ListAssessmentRunAgentsResponse, ListAssessmentRunAgentsError>;

    fn list_assessment_runs(
        &self, 
        input: ListAssessmentRunsRequest
    ) -> RusotoFuture<ListAssessmentRunsResponse, ListAssessmentRunsError>;

    fn list_assessment_targets(
        &self, 
        input: ListAssessmentTargetsRequest
    ) -> RusotoFuture<ListAssessmentTargetsResponse, ListAssessmentTargetsError>;

    fn list_assessment_templates(
        &self, 
        input: ListAssessmentTemplatesRequest
    ) -> RusotoFuture<ListAssessmentTemplatesResponse, ListAssessmentTemplatesError>;

    fn list_event_subscriptions(
        &self, 
        input: ListEventSubscriptionsRequest
    ) -> RusotoFuture<ListEventSubscriptionsResponse, ListEventSubscriptionsError>;

    fn list_exclusions(
        &self, 
        input: ListExclusionsRequest
    ) -> RusotoFuture<ListExclusionsResponse, ListExclusionsError>;

    fn list_findings(
        &self, 
        input: ListFindingsRequest
    ) -> RusotoFuture<ListFindingsResponse, ListFindingsError>;

    fn list_rules_packages(
        &self, 
        input: ListRulesPackagesRequest
    ) -> RusotoFuture<ListRulesPackagesResponse, ListRulesPackagesError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn preview_agents(
        &self, 
        input: PreviewAgentsRequest
    ) -> RusotoFuture<PreviewAgentsResponse, PreviewAgentsError>;

    fn register_cross_account_access_role(
        &self, 
        input: RegisterCrossAccountAccessRoleRequest
    ) -> RusotoFuture<(), RegisterCrossAccountAccessRoleError>;

    fn remove_attributes_from_findings(
        &self, 
        input: RemoveAttributesFromFindingsRequest
    ) -> RusotoFuture<RemoveAttributesFromFindingsResponse, RemoveAttributesFromFindingsError>;

    fn set_tags_for_resource(
        &self, 
        input: SetTagsForResourceRequest
    ) -> RusotoFuture<(), SetTagsForResourceError>;

    fn start_assessment_run(
        &self, 
        input: StartAssessmentRunRequest
    ) -> RusotoFuture<StartAssessmentRunResponse, StartAssessmentRunError>;

    fn stop_assessment_run(
        &self, 
        input: StopAssessmentRunRequest
    ) -> RusotoFuture<(), StopAssessmentRunError>;

    fn subscribe_to_event(
        &self, 
        input: SubscribeToEventRequest
    ) -> RusotoFuture<(), SubscribeToEventError>;

    fn unsubscribe_from_event(
        &self, 
        input: UnsubscribeFromEventRequest
    ) -> RusotoFuture<(), UnsubscribeFromEventError>;

    fn update_assessment_target(
        &self, 
        input: UpdateAssessmentTargetRequest
    ) -> RusotoFuture<(), UpdateAssessmentTargetError>;
}

Trait representing the capabilities of the Amazon Inspector API. Amazon Inspector clients implement this trait.

Required methods

fn add_attributes_to_findings(
&self,
input: AddAttributesToFindingsRequest
) -> RusotoFuture<AddAttributesToFindingsResponse, AddAttributesToFindingsError>

Assigns attributes (key and value pairs) to the findings that are specified by the ARNs of the findings.

fn create_assessment_target(
&self,
input: CreateAssessmentTargetRequest
) -> RusotoFuture<CreateAssessmentTargetResponse, CreateAssessmentTargetError>

Creates a new assessment target using the ARN of the resource group that is generated by CreateResourceGroup. If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target. If the service-linked role isn’t already registered, this action also creates and registers a service-linked role to grant Amazon Inspector access to AWS Services needed to perform security assessments. You can create up to 50 assessment targets per AWS account. You can run up to 500 concurrent agents per AWS account. For more information, see Amazon Inspector Assessment Targets.

fn create_assessment_template(
&self,
input: CreateAssessmentTemplateRequest
) -> RusotoFuture<CreateAssessmentTemplateResponse, CreateAssessmentTemplateError>

Creates an assessment template for the assessment target that is specified by the ARN of the assessment target. If the service-linked role isn’t already registered, this action also creates and registers a service-linked role to grant Amazon Inspector access to AWS Services needed to perform security assessments.

fn create_exclusions_preview(
&self,
input: CreateExclusionsPreviewRequest
) -> RusotoFuture<CreateExclusionsPreviewResponse, CreateExclusionsPreviewError>

Starts the generation of an exclusions preview for the specified assessment template. The exclusions preview lists the potential exclusions (ExclusionPreview) that Inspector can detect before it runs the assessment.

fn create_resource_group(
&self,
input: CreateResourceGroupRequest
) -> RusotoFuture<CreateResourceGroupResponse, CreateResourceGroupError>

Creates a resource group using the specified set of tags (key and value pairs) that are used to select the EC2 instances to be included in an Amazon Inspector assessment target. The created resource group is then used to create an Amazon Inspector assessment target. For more information, see CreateAssessmentTarget.

fn delete_assessment_run(
&self,
input: DeleteAssessmentRunRequest
) -> RusotoFuture<(), DeleteAssessmentRunError>

Deletes the assessment run that is specified by the ARN of the assessment run.

fn delete_assessment_target(
&self,
input: DeleteAssessmentTargetRequest
) -> RusotoFuture<(), DeleteAssessmentTargetError>

Deletes the assessment target that is specified by the ARN of the assessment target.

fn delete_assessment_template(
&self,
input: DeleteAssessmentTemplateRequest
) -> RusotoFuture<(), DeleteAssessmentTemplateError>

Deletes the assessment template that is specified by the ARN of the assessment template.

fn describe_assessment_runs(
&self,
input: DescribeAssessmentRunsRequest
) -> RusotoFuture<DescribeAssessmentRunsResponse, DescribeAssessmentRunsError>

Describes the assessment runs that are specified by the ARNs of the assessment runs.

fn describe_assessment_targets(
&self,
input: DescribeAssessmentTargetsRequest
) -> RusotoFuture<DescribeAssessmentTargetsResponse, DescribeAssessmentTargetsError>

Describes the assessment targets that are specified by the ARNs of the assessment targets.

fn describe_assessment_templates(
&self,
input: DescribeAssessmentTemplatesRequest
) -> RusotoFuture<DescribeAssessmentTemplatesResponse, DescribeAssessmentTemplatesError>

Describes the assessment templates that are specified by the ARNs of the assessment templates.

fn describe_cross_account_access_role(
&self
) -> RusotoFuture<DescribeCrossAccountAccessRoleResponse, DescribeCrossAccountAccessRoleError>

Describes the IAM role that enables Amazon Inspector to access your AWS account.

fn describe_exclusions(
&self,
input: DescribeExclusionsRequest
) -> RusotoFuture<DescribeExclusionsResponse, DescribeExclusionsError>

Describes the exclusions that are specified by the exclusions' ARNs.

fn describe_findings(
&self,
input: DescribeFindingsRequest
) -> RusotoFuture<DescribeFindingsResponse, DescribeFindingsError>

Describes the findings that are specified by the ARNs of the findings.

fn describe_resource_groups(
&self,
input: DescribeResourceGroupsRequest
) -> RusotoFuture<DescribeResourceGroupsResponse, DescribeResourceGroupsError>

Describes the resource groups that are specified by the ARNs of the resource groups.

fn describe_rules_packages(
&self,
input: DescribeRulesPackagesRequest
) -> RusotoFuture<DescribeRulesPackagesResponse, DescribeRulesPackagesError>

Describes the rules packages that are specified by the ARNs of the rules packages.

fn get_assessment_report(
&self,
input: GetAssessmentReportRequest
) -> RusotoFuture<GetAssessmentReportResponse, GetAssessmentReportError>

Produces an assessment report that includes detailed and comprehensive results of a specified assessment run.

fn get_exclusions_preview(
&self,
input: GetExclusionsPreviewRequest
) -> RusotoFuture<GetExclusionsPreviewResponse, GetExclusionsPreviewError>

Retrieves the exclusions preview (a list of ExclusionPreview objects) specified by the preview token. You can obtain the preview token by running the CreateExclusionsPreview API.

fn get_telemetry_metadata(
&self,
input: GetTelemetryMetadataRequest
) -> RusotoFuture<GetTelemetryMetadataResponse, GetTelemetryMetadataError>

Information about the data that is collected for the specified assessment run.

fn list_assessment_run_agents(
&self,
input: ListAssessmentRunAgentsRequest
) -> RusotoFuture<ListAssessmentRunAgentsResponse, ListAssessmentRunAgentsError>

Lists the agents of the assessment runs that are specified by the ARNs of the assessment runs.

fn list_assessment_runs(
&self,
input: ListAssessmentRunsRequest
) -> RusotoFuture<ListAssessmentRunsResponse, ListAssessmentRunsError>

Lists the assessment runs that correspond to the assessment templates that are specified by the ARNs of the assessment templates.

fn list_assessment_targets(
&self,
input: ListAssessmentTargetsRequest
) -> RusotoFuture<ListAssessmentTargetsResponse, ListAssessmentTargetsError>

Lists the ARNs of the assessment targets within this AWS account. For more information about assessment targets, see Amazon Inspector Assessment Targets.

fn list_assessment_templates(
&self,
input: ListAssessmentTemplatesRequest
) -> RusotoFuture<ListAssessmentTemplatesResponse, ListAssessmentTemplatesError>

Lists the assessment templates that correspond to the assessment targets that are specified by the ARNs of the assessment targets.

fn list_event_subscriptions(
&self,
input: ListEventSubscriptionsRequest
) -> RusotoFuture<ListEventSubscriptionsResponse, ListEventSubscriptionsError>

Lists all the event subscriptions for the assessment template that is specified by the ARN of the assessment template. For more information, see SubscribeToEvent and UnsubscribeFromEvent.

fn list_exclusions(
&self,
input: ListExclusionsRequest
) -> RusotoFuture<ListExclusionsResponse, ListExclusionsError>

List exclusions that are generated by the assessment run.

fn list_findings(
&self,
input: ListFindingsRequest
) -> RusotoFuture<ListFindingsResponse, ListFindingsError>

Lists findings that are generated by the assessment runs that are specified by the ARNs of the assessment runs.

fn list_rules_packages(
&self,
input: ListRulesPackagesRequest
) -> RusotoFuture<ListRulesPackagesResponse, ListRulesPackagesError>

Lists all available Amazon Inspector rules packages.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Lists all tags associated with an assessment template.

fn preview_agents(
&self,
input: PreviewAgentsRequest
) -> RusotoFuture<PreviewAgentsResponse, PreviewAgentsError>

Previews the agents installed on the EC2 instances that are part of the specified assessment target.

fn register_cross_account_access_role(
&self,
input: RegisterCrossAccountAccessRoleRequest
) -> RusotoFuture<(), RegisterCrossAccountAccessRoleError>

Registers the IAM role that grants Amazon Inspector access to AWS Services needed to perform security assessments.

fn remove_attributes_from_findings(
&self,
input: RemoveAttributesFromFindingsRequest
) -> RusotoFuture<RemoveAttributesFromFindingsResponse, RemoveAttributesFromFindingsError>

Removes entire attributes (key and value pairs) from the findings that are specified by the ARNs of the findings where an attribute with the specified key exists.

fn set_tags_for_resource(
&self,
input: SetTagsForResourceRequest
) -> RusotoFuture<(), SetTagsForResourceError>

Sets tags (key and value pairs) to the assessment template that is specified by the ARN of the assessment template.

fn start_assessment_run(
&self,
input: StartAssessmentRunRequest
) -> RusotoFuture<StartAssessmentRunResponse, StartAssessmentRunError>

Starts the assessment run specified by the ARN of the assessment template. For this API to function properly, you must not exceed the limit of running up to 500 concurrent agents per AWS account.

fn stop_assessment_run(
&self,
input: StopAssessmentRunRequest
) -> RusotoFuture<(), StopAssessmentRunError>

Stops the assessment run that is specified by the ARN of the assessment run.

fn subscribe_to_event(
&self,
input: SubscribeToEventRequest
) -> RusotoFuture<(), SubscribeToEventError>

Enables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

fn unsubscribe_from_event(
&self,
input: UnsubscribeFromEventRequest
) -> RusotoFuture<(), UnsubscribeFromEventError>

Disables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

fn update_assessment_target(
&self,
input: UpdateAssessmentTargetRequest
) -> RusotoFuture<(), UpdateAssessmentTargetError>

Updates the assessment target that is specified by the ARN of the assessment target.

If resourceGroupArn is not specified, all EC2 instances in the current AWS account and region are included in the assessment target.

Loading content...